requirements.txt
fastapi
uvicorn[standard]
playwright
apscheduler
asyncpg
python-dotenv
